\name{StartupMessage}
\alias{StartupMessage}
\alias{gammiStartupMessage}

\title{Startup Message for gammi}

\description{
Prints the startup message when gammi is loaded. Not intended to be called by the user.
}

\details{
The `gammi' ascii start-up message was created using the taag software.
}

\references{
https://patorjk.com/software/taag/
}